Turn Order + Several other changes
In the mainClass, I added a method to determine turn order that is
pretty modular, as well as a bunch of test prints. In the playerUnit, I
added several update functions and a block charging method. In the
Equipment, I just added a small test case. In the statTree, I
implemented what we currently have for the skill tree into the code.

Overall I added a bunch of comments and small formatting changes.